+++ admin/drivers/pgsql.inc.php
@@ -383,7 +383,10 @@ if (isset($_GET["pgsql"])) {
 			}
 
 			if ($connection->isMinVersion("12")) {
-				$this->generated = ["STORED"];
+				$this->generated[] = "STORED";
+				if ($connection->isMinVersion("18")) {
+					$this->generated[] = "VIRTUAL";
+				}
 			}
 
 			// No "SQL" to avoid CSRF.
@@ -843,7 +846,8 @@ ORDER BY a.attnum"
 			if (in_array($row['attidentity'], ['a', 'd'])) {
 				$row['default'] = 'GENERATED ' . ($row['attidentity'] == 'd' ? 'BY DEFAULT' : 'ALWAYS') . ' AS IDENTITY';
 			}
-			$row["generated"] = ($row["attgenerated"] == "s" ? "STORED" : "");
+			$options = ["s" => "STORED", "v" => "VIRTUAL"];
+			$row["generated"] = ($options[$row["attgenerated"]] ?? "");
 			$row["null"] = !$row["attnotnull"];
 			$row["auto_increment"] = $row['attidentity'] || preg_match('~^nextval\(~i', $row["default"])
 				|| preg_match('~^unique_rowid\(~', $row["default"]); // CockroachDB
@@ -1010,7 +1014,7 @@ ORDER BY s.ordinal_position";
 					}
 					$alter[] = "ALTER $column TYPE$val[1]";
 					$sequence_name = $table . "_" . idf_unescape($val[0]) . "_seq";
-					$alter[] = "ALTER $column " . ($val[3] ? "SET" . preg_replace('~GENERATED ALWAYS(.*) STORED~', 'EXPRESSION\1', $val[3])
+					$alter[] = "ALTER $column " . ($val[3] ? "SET" . preg_replace('~GENERATED ALWAYS(.*) (STORED|VIRTUAL)~', 'EXPRESSION\1', $val[3])
 						: (isset($val[6]) ? "SET DEFAULT nextval(" . q($sequence_name) . ")"
 						: "DROP DEFAULT" //! change to DROP EXPRESSION with generated columns
 					));